backend/requirements.txt
# Backend Dependencies

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0

# Database
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
pymongo>=4.10.1

# AI/ML
openai>=1.59.0
langchain>=0.3.0
langchain-openai>=0.2.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
schedule==1.2.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Date/Time
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Logging
loguru==0.7.2
